Move Reference Types to their own file
